Temporary bugfix
Instead of opening a new tab, the script will report a string to the
Macro Window. This is due to frequent hangs caused by the function that
tried to open a new tab.

If this was not a problem for you, feel free to remove the pound sign
and the space at the beginning of line 24.